% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lga-sga.R
\name{ga_class}
\alias{ga_class}
\title{Calculate a single percentile class of birthweight}
\usage{
ga_class(ga, bsex, birthweight)
}
\arguments{
\item{ga}{An integer representing a neonate's gestational age in weeks}

\item{bsex}{A character indicating sex ("M"/"F")}

\item{birthweight}{An integer representing birthweight in grams}
}
\value{
A string representing the percentile class of birthweight given gestational age and sex
}
\examples{
ga_class(ga = 25, bsex = "M", birthweight = 2521)
}
