% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lga-sga.R
\name{ga_dataframe}
\alias{ga_dataframe}
\title{Calculate a percentile class of birthweight over a dataframe}
\usage{
ga_dataframe(DF, gestation_age = "combgest", neonate_sex = "sex",
  birthweight = "dbwt")
}
\arguments{
\item{DF}{A dataframe containing columns for (at least) gestational age, sex, and birthweight}

\item{gestation_age}{A column of integers representing neonates' gestational age in weeks}

\item{neonate_sex}{A column of characters indicating sex ("M"/"F")}

\item{birthweight}{A column of integers representing birthweight in grams}
}
\value{
DF with an additional column (gest_cat) representingthe percentile class of birthweight
}
\examples{
ga_dataframe(A_DataFrame)
ga_dataframe(A_DataFrame, gestation_age = "my_gest", neonate_sex = "my_sex", birthweight = "my_bwt"
}
